refactor: update python scene generator
