+++ scripts/pyproject.toml
@@ -1,0 +1,17 @@
+[project]
+name = "scripts"
+version = "0.1.0"
+description = "Scripts to generate scenes"
+readme = "README.md"
+authors = [
+    { name = "Julien Graziano", email = "julien@graziano.fr" }
+]
+requires-python = ">=3.13"
+dependencies = []
+
+[project.scripts]
+generate = "scripts:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
