+++ scripts/src/scripts/common/seq.py
@@ -1,0 +1,5 @@
+def seq(start: float, stop: float, step: float):
+    v = start
+    while v <= stop:
+        yield v
+        v += step
