+++ scripts/src/scripts/common/serializable.py
@@ -1,0 +1,2 @@
+class Serializable:
+    def serialize(self) -> dict | float | int | list | str | tuple: ...
